+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 All notable changes to the "vsc-bigquery" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
+### Added
+
 - Initial release
+- Basic syntax highlighting (keywords, strings, comments)
+- Dry runs after editing text
+- Estimated processed bytes in status bar
+- Listing of accessible projects, datasets, tables
+- Listing of tables schemas
+- Formatting using forked sql-formatter